feat: add export map
