+++ packages/di/dom/lib/environment.ts
